fix logic error in ec2 get_instance_userdata and slow tests
The change to get_instance_userdata is to fix an issue that
was causing retry in the test when it was not desired.
if user_data returned 404 it means "there was no user-data", so
dont bother retrying.  However, _skip_retry_on_codes was returning
False indicating that readurl should retry.

test_merging was creating 2500 random tests, shrink that down to 100.
test_seed_runs is still on my system the slowest test, but 
taking < .5 seconds where it was taking > 3.