Allow Azure provisioning without WALinuxAgent.
We now leverage the ability for the dhcp client to log
its known information on exit of the client.  This is done
in different ways depending on the networking stack, such
as natively vs NetworkManager.  We now look for the special
DHCP option 245 in the options first and then fall back to
examining the log files.